refactor: Export db package
To allow external applications the ability to reuse the db Prepare and Teardown functions.
